moved the physics code to its own source file
